services/chatbot_service/requirements.txt
Django>=3.2,<4.3
djangorestframework>=3.12,<3.15
django-cors-headers>=3.0,<4.0